requirement.txt
flask
flask-mysqldb
flask-bcrypt
python-dotenv

